ci: Test against current and previous Go versions
Go's support policy is to support the two most recent minor releases:
e.g., with the release of Go 1.25, the supported versions are
1.24 and 1.25. Most libraries follow the same support policy.

While the support policy for autogold is not necessarily written down,
given that 1.24 and 1.25 (current stable and oldstable) work fine,
test against both in CI until there's a reason not to.

Note:
The "stable" and "olstable" aliases require setup-go v3 or later.